requirements.txt
fastapi
pydantic
uvicorn
xgboost
pandas
numpy
scikit-learn
python-dotenv
httpx
pytest
ipykernel
coverage
openpyxl
matplotlib
scipy